Adds video & artist name content to /catalogue
I added functions that enable the pause/play functionality
to the video when mouseover/mouseleave happens, respectively.
However, the video is not showing up, so I will be working
to debug that.

Also, the artist names are showing up, though not spaced
correctly. This is probably just a stylistic issue though.
TBD.